partie API un chouia avancée, a voir avec Olivier pour le truc des id -> DTO

+++ swagger/spring-server/src/main/java/io/pestakit/status/entities/ServiceEntity.java
@@ -11,7 +11,7 @@ public class ServiceEntity implements Serializable
 {
    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
-   private long id;
+   private Integer id;
 
    private String name;
    private String statusAddress;
@@ -19,12 +19,12 @@ public class ServiceEntity implements Serializable
    private String state;
    private String contact;
 
-   public long getId()
+   public Integer getId()
    {
       return id;
    }
 
-   public void setId(long id)
+   public void setId(Integer id)
    {
       this.id = id;
    }
